Create README.md
This is site is an endpoint for automated builds of collections such as `TS-Kata-Collections` and `TS-VR-Katas`.  These build processes dump files into `GE` and `VR`, respectively.

The only part that you may need to edit by hand is the `build-index.rkt` script, which generates the top-level `index.html` file when run.  Edit that script, run it, and push it to master to see changes on the landing page.